chore(CI): fix GitHub `set-output` deprecation warning
