+++ .github/workflows/test-macOS.yml
@@ -35,18 +35,20 @@ jobs:
       - name: Install Pnpm
         run: corepack enable
 
-      - name: Check docs only change
-        run: echo "::set-output name=DOCS_CHANGE::$(node ./scripts/skipDocsChange.js echo 'not-docs-only-change')"
-        id: docs-change
+      - name: Check skip CI
+        run: echo "RESULT=$(node ./scripts/skipCI.js)" >> "$GITHUB_OUTPUT"
+        id: skip-ci
 
-      - run: echo "${{steps.docs-change.outputs.DOCS_CHANGE}}"
+      - name: Log skip CI result
+        run: echo "${{steps.skip-ci.outputs.RESULT}}"
 
       - name: Setup Node.js ${{ matrix.node-version }}
-        if: ${{steps.docs-change.outputs.DOCS_CHANGE != 'docs only change'}}
+        if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
         uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
           cache: 'pnpm'
+
       - name: Turbo Cache
         id: turbo-cache
         uses: actions/cache@v3
@@ -58,13 +60,13 @@ jobs:
             turbo-
 
       - name: Install Dependencies
-        if: ${{steps.docs-change.outputs.DOCS_CHANGE != 'docs only change'}}
+        if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
         run: pnpm install --ignore-scripts
 
       - name: Unit Test
-        if: ${{steps.docs-change.outputs.DOCS_CHANGE != 'docs only change'}}
+        if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
         run: pnpm run test:unit
 
       - name: E2E Test
-        if: ${{steps.docs-change.outputs.DOCS_CHANGE != 'docs only change'}}
+        if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
         run: npx playwright install && pnpm run test:e2e
